Change argument type in -locationOfLastWhitespaceLessThan: Added setting of typing attributes for command words.

+++ TeXnicle/TeXEditor/TextView/TeXTextView.h
@@ -140,7 +140,7 @@
 
 - (NSInteger) lengthOfLineUpToLocation:(NSUInteger)location;
 - (NSRange) rangeForCurrentParagraph;
-- (NSInteger) locationOfLastWhitespaceLessThan:(int)lineWrapLength;
+- (NSInteger) locationOfLastWhitespaceLessThan:(NSInteger)lineWrapLength;
 - (NSString*) currentLineToCursor;
 - (NSRange) getVisibleRange;
 - (void) selectUpToCurrentLocation;
